Added Home.html and style_home.css
I have deleted the home page branch and renamed this one. Since I will be working on both home and login page, I will be working on this branch itself. In addition, adding an HTML and a CSS file for the home page. This page contains buttons which redirect to the login page and various other information. Some button functionality is yet to be programmed. For now, I have just linked the home and login/registration pages using this code.